// src/app/openapi/model/bewertung.ts
/**
 * Filmkritiken-Backend
 * Filmkritiken-Backend
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Bewertung { 
    von: string;
    wertung: number;
    /**
     * Wenn gesetzt, gilt die Bewertung als aktive Enthaltung.
     */
    enthaltung?: boolean;
}